split the controllers files relates #17
