--- server/controller/Authentication/checkNewUserData.js
@@ -1,6 +1,0 @@
-const { checkNewUserDataQ } = require("../../database/queries");
-
-const checkNewUserData = (email, username) => {
-  return checkNewUserDataQ(email, username);
-};
-module.exports = { checkNewUserData };
