+++ server/controller/Authentication/index.js
@@ -1,0 +1,6 @@
+const generateToken = require("./generateToken");
+const login = require("./login");
+const logout = require("./logout");
+const signUp = require("./signUp");
+
+module.exports = { generateToken, login, logout, signUp };
