+++ server/controller/comments/index.js
@@ -1,0 +1,5 @@
+const postNewComment = require("./postNewComment");
+const getPostComments = require("./getPostComments");
+const deleteComment = require("./deleteComment");
+
+module.exports = { postNewComment, getPostComments, deleteComment };
