+++ server/controller/index.js
@@ -1,13 +1,6 @@
-const signUp = require("./Authentication/signUp");
-const login = require("./Authentication/login");
-const logout = require("./Authentication/logout");
-const searchPost = require("./posts/searchPost");
-const getAllPosts = require("./posts/getAllPosts");
-const addNewPost = require("./posts/addNewPost");
-const getPostComment = require("./comments/getPostComments");
-const postNewComment = require("./comments/postNewComment");
-const deleteComment = require("./comments/deleteComment");
-const deletePost = require("./posts/deletePost");
+const { login, logout, signUp } = require("./Authentication");
+const { postNewComment, getPostComments, deleteComment } = require("./comments");
+const { searchPost, getAllPosts, deletePost, addNewPost } = require("./posts");
 
 module.exports = {
   signUp,
@@ -15,7 +8,7 @@ module.exports = {
   logout,
   searchPost,
   getAllPosts,
-  getPostComment,
+  getPostComments,
   postNewComment,
   addNewPost,
   deleteComment,
