+++ server/controller/posts/index.js
@@ -1,0 +1,6 @@
+const searchPost = require("./searchPost");
+const getAllPosts = require("./getAllPosts");
+const deletePost = require("./deletePost");
+const addNewPost = require("./addNewPost");
+
+module.exports = { searchPost, getAllPosts, deletePost, addNewPost };
