+++ server/validation/index.js
@@ -1,0 +1,4 @@
+const signupValidation = require("./signupValidation");
+const loginValidation = require("./loginValidation");
+
+module.exports = { signupValidation, loginValidation };
